Hacky addition of visible-only row rendering and fixed table height
For tables with lots of data (hundreds to thousands of rows), rendering all
rows is very slow. This change adds an option to only render visible rows
(requires the table to be scrollable - fixed height option allows this),
which significantly improves performance.

Margins are added to the table element at the top and bottom to simulate
all rows being present in the container (allowing full scroll), while only
the rows calculated to be in the visible part of the window are rendered.

A consistent row height is assumed (with a hard-coded default) - rows with
varying heights will break scrolling and cause unexpected behavior.